fastapi-notes/25-chitravarnan/requirements.txt
fastapi[standard]>=0.110.0
uvicorn>=0.27.0
google-generativeai>=0.4.0
pillow>=10.0.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
